feat: Fuata class initialisation
The Fuata class will contain all methods that will be invoked when
certain commands are provided in the terminal. When `fuata init` is
entered, a Fuata object will be created subsequently runnning the init
block. Additionally, unit tests have been written to test when the
command is run for the first time within a directory and also when it is
run for the second time