remove: Decompression of data read from index file
Previously, Indexing.addToFileIndex() would decompress the data existing
in the index file after reading it on line 41. This was causing an
error while reading the file since it was presented in the wrong format.
To correct this, the content was simply read using the
Files.readString() method that reads the Bytes from the file and returns
a CharSequence; a string literal