Update: TreeUtil.getDirectoryPath to use '\'
Previously, the function used the forwardslash charachter to identify
directory paths. This would work on a Unix system. However, on a Windows
OS, directory paths are separated using a backslash character. This was
causing undefined behaviour. To solve this, line 84 of TreeUtil.kt was
changed to "\\"; the first backslash being an escape character.
Possibly, this line would need to be changed in the future to accomodate
both Unix and Windows OSs