Update: move initRepository from the init block
Previously, the .fuata repository initialisation was placed in the Fuata
class constructor. This architecture would bring a problem when a Fuata
instance would need to be created to access the other class functions.
An exception would be thrown everytime a Fuata object is created in the
case where the .fuata repository proper already exists, causing the
program to crash. This has been changed by moving the initialisation
logic to its own method; initialiseRepository. Additionally, functions
to check whether the repository proper exists have been added