Add: tasks.jar manifest in build.gradle
This addition makes it possible to locate the entry point for the
package. Additionally, the ShadowJar dependency has been added to enable
the fat JAR when building the project. A fat JAR enables the inclusion
of dependencies and libraries in the final build output so that no
additional setup is needed