fix: making multiple commits feature
Previously, when making a multiple commits - commits after making the
first one - the operation would not be successful because there was a
bug in reading the tree object file to retrieve the root tree of the
parent commit. The function carrying out this particular operation is
TreeUtil.getTree(). The issue was on line 109 of TreeUtil.kt. Whereas
the function is supposed to return a Tree wrapped in a result, it was
returning a Tree, causing an exception to be thrown. Since this function
is invoked in CommitUtil.createCommit(), it would result in the
operation breaking. This bug has been corrected by adding a Result
wrapper to "Json.decodeFromString(treeJson)" on line 109