fix: writing of content in index file
Previously, Indexing.addFileToIndex() was writing a compressed ByteArray in the index file after creating the JSON array of IndexEntry objects. This was causing an error when staging multiple files for a single commit because there was an error in reading the file content. Whereas bytes were expected to be read and a CharSequence returned, it was finding a compressed ByteArray. This mistake has been corrected on line 51 of Indexing.kt